feat: 설계에 따라 프론트엔드 디렉토리 구조 수정
