+++ .gitignore
@@ -28,7 +28,10 @@ backend/.gradle/
 backend/build/
 backend/application-secret.properties
 backend/application.secret.yml
+application-secret.properties
+application.secret.yml
 backend/application.log*
+application.log*
 
 # IDE 및 빌드 관련 파일
 backend/.classpath
