+++ frontend/src/assets/images/favicon.png
