+++ frontend/src/assets/images/react-logo@3x.png
