+++ frontend/src/assets/images/splash-icon.png
