+++ frontend/src/components/external-link.tsx
