+++ frontend/src/components/haptic-tab.tsx
