+++ frontend/src/components/ui/icon-symbol.tsx
