+++ frontend/src/scripts/reset-project.js
